requirements.txt
huggingface_hub>=0.23.0
llama-cpp-python>=0.2.83
safetensors>=0.4.2
sentencepiece>=0.2.0
transformers>=4.45.0
torch>=2.2.0
